c57

+++ common/rust/cctrusted_base/src/eventlog.rs
@@ -75,7 +75,7 @@ impl TcgEventLog {
         //Ok(self.event_logs[begin as usize..end as usize].to_vec())
         let return_event_logs: Vec<EventLogEntryType> = Vec::new();
         for idx in begin..end {
-            return_event_logs.push(self.event_logs[idx]);
+            return_event_logs.push(self.event_logs[idx as usize]);
         }
         Ok(return_event_logs)
     }
